+++ federation/src/main/scala/caliban/federation/package.scala
@@ -13,7 +13,7 @@ import caliban.federation.v2x.{
 
 package object federation {
 
-  lazy val v1    = new FederationV1 with FederationDirectives
+  lazy val v1    = new FederationV1 with FederationDirectives with FederationDirectivesV1
   lazy val v2_0  = new FederationV2(List(Versions.v2_0))
   lazy val v2_1  = new FederationV2(List(Versions.v2_1))
   lazy val v2_3  = new FederationV2(List(Versions.v2_3)) with FederationDirectivesV2_3
