+++ federation/src/main/scala/caliban/federation/v2x/FederationDirectivesV2.scala
@@ -1,6 +1,7 @@
 package caliban.federation.v2x
 
-import caliban.Value.StringValue
+import caliban.InputValue
+import caliban.Value.{ BooleanValue, StringValue }
 import caliban.parsing.adt.Directive
 import caliban.schema.Annotations.GQLDirective
 
@@ -25,6 +26,19 @@ trait FederationDirectivesV2 {
   object Tag {
     def apply(name: String): Directive =
       Directive("tag", Map("name" -> StringValue(name)))
+  }
+
+  case class GQLKey(fields: String, resolvable: Boolean = true) extends GQLDirective(Key(fields, resolvable))
+
+  object Key {
+    def apply(fields: String, resolvable: Boolean = true, name: String = "key"): Directive = {
+      val args: List[(String, InputValue)] = List(
+        Some("fields" -> StringValue(fields)),
+        (if (resolvable) None else Some("resolvable" -> BooleanValue(resolvable)))
+      ).flatten
+
+      Directive(name, args.toMap)
+    }
   }
 
 }
